# api-specs.yaml
openapi: 3.0.3
info:
  title: Clustr Platform API
  version: 1.0.0
  description: API specification for Clustr.
components:
  schemas:
...
    UserSummary:
      type: object
      description: A summary of user information, typically for lists (e.g., Manage Users, Chat participants).
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user.
        fullName:
          type: string
          description: Combined first and last name of the user.
          example: "Maimuna Lawal"
        profilePictureUrl:
          type: string
          format: uri
          description: URL to the user's profile picture.

    # --- Home Screen Schemas ---
...
    PasswordChange:
      type: object
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
        confirmPassword:
          type: string
          format: password
      required:
        - oldPassword
        - newPassword
        - confirmPassword
...
    PaginatedResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          description: The array of items for the current page.
          items: {} # Can be any type
        pagination:
          type: object
          required:
            - page
            - limit
            - total
            - totalPages
          properties:
            page:
              type: integer
              minimum: 1
              description: The current page number.
            limit:
              type: integer
              minimum: 1
              maximum: 100
              description: The number of items per page.
            total:
              type: integer
              minimum: 0
              description: The total number of items across all pages.
            totalPages:
              type: integer
              minimum: 0
              description: The total number of pages.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication.
paths:
...
  /users/search:
    get:
      tags: [Users]
      summary: Search users by name or email.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            description: Search query for user name or email.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserSummary'
        '401':
          description: Unauthorized.

  # --- Group Paths ---
...
  /users/{userId}/change-password:
    post:
      summary: Change user password
      tags:
        - User Management
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully"
...